Flask - API/Plant_disease_predictor/requirements.txt
numpy
flask
opencv-python
tensorflow
keras
pillow
